chore: store for in-app module (#430)
