+++ core/src/main/kotlin/io/customer/sdk/data/store/GlobalPreferenceStore.kt
@@ -10,10 +10,7 @@ import androidx.core.content.edit
 interface GlobalPreferenceStore {
     fun saveDeviceToken(token: String)
     fun getDeviceToken(): String?
-    fun saveUserId(token: String)
-    fun getUserId(): String?
     fun removeDeviceToken()
-    fun removeUserId()
     fun clear(key: String)
     fun clearAll()
 }
@@ -34,20 +31,9 @@ internal class GlobalPreferenceStoreImpl(
         getString(KEY_DEVICE_TOKEN, null)
     }
 
-    override fun saveUserId(token: String) = prefs.edit {
-        putString(KEY_USER_ID, token)
-    }
-
-    override fun getUserId(): String? = prefs.read {
-        getString(KEY_USER_ID, null)
-    }
-
     override fun removeDeviceToken() = clear(KEY_DEVICE_TOKEN)
-
-    override fun removeUserId() = clear(KEY_USER_ID)
 
     companion object {
         private const val KEY_DEVICE_TOKEN = "device_token"
-        private const val KEY_USER_ID = "user_id"
     }
 }
